pyproject.toml
[project]
name = "hackthon-api"
version = "1.0.0"
authors = [
    {name = "Conor McKillop", email = "contact@cmckillop.dev"},
]
dependencies = [
    "fastapi>=0.103.2",
    "uvicorn[standard]>=0.23.2",
    "black>=23.9.1",
    "ruff>=0.0.292",
    "sqlmodel>=0.0.8",
    "psycopg2>=2.9.9",
    "pyjwt>=2.8.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.poetry]
name = "hackthon-api"
version = "1.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.11"